split form code into divided renderers
