+++ src/form.tsx
@@ -1,21 +1,16 @@
 import React, { SFC, ReactNode, CSSProperties, useState, useEffect } from "react";
 import { row, column, flex } from "@jimengio/shared-utils";
 import { css, cx } from "emotion";
-import Input from "antd/lib/input";
-import Switch from "antd/lib/switch";
-import Select from "antd/lib/select";
-import InputNumber from "antd/lib/input-number";
-import { lingual, formatString } from "./lingual";
 import { IMesonFieldItem, EMesonFieldType, IMesonFieldItemHasValue, FuncMesonModifyForm, IMesonErrors, IMesonFormBase } from "./model/types";
 
 import { RequiredMark } from "./component/misc";
 import { FormFooter, EMesonFooterLayout } from "./component/form-footer";
 import MesonModal from "./component/modal";
-import TextArea from "antd/lib/input/TextArea";
 import produce, { Draft } from "immer";
 import MesonDrawer from "./component/drawer";
 import { useMesonCore } from "./hook/meson-core";
 import { showErrorByNames } from "./util/validation";
+import { renderTextAreaItem, renderInputItem, renderNumberItem, renderSelectItem, renderSwitchItem, renderItemLayout } from "./renderer";
 
 /**
  * 清空draft对象的value值
@@ -92,120 +87,16 @@ export function ForwardForm<T = IMesonFormBase>(props: MesonFormProps<T>, ref: R
     switch (item.type) {
       case EMesonFieldType.Input:
         if (item.textarea) {
-          return (
-            <div className={cx(styleControlBase, styleTextareaBase)}>
-              <TextArea
-                value={form[item.name]}
-                disabled={item.disabled}
-                placeholder={item.placeholder || formatString(lingual.pleaseInputLabel, { label: item.label })}
-                onChange={(event) => {
-                  let newValue = event.target.value;
-                  updateItem(newValue, item);
-                }}
-                onBlur={(event: any) => {
-                  checkItem(item);
-                }}
-                // should use TextareaProps, but for convenience
-                {...item.inputProps as any}
-              />
-            </div>
-          );
+          return renderTextAreaItem(form, item, updateItem, checkItem);
         }
-        return (
-          <div className={styleControlBase}>
-            <Input
-              value={form[item.name]}
-              disabled={item.disabled}
-              type={item.inputType || "text"}
-              placeholder={item.placeholder || formatString(lingual.pleaseInputLabel, { label: item.label })}
-              onChange={(event) => {
-                let newValue = event.target.value;
+        return renderInputItem(form, item, updateItem, checkItem);
 
-                // reset empty string to undefined by default, FR-96
-                if (newValue.trim() === "") {
-                  if (!item.useBlank) {
-                    newValue = undefined;
-                  }
-                }
-
-                updateItem(newValue, item);
-              }}
-              onBlur={() => {
-                checkItem(item);
-              }}
-              {...item.inputProps}
-            />
-          </div>
-        );
       case EMesonFieldType.Number:
-        return (
-          <div className={styleControlBase}>
-            <InputNumber
-              value={form[item.name]}
-              disabled={item.disabled}
-              placeholder={item.placeholder || formatString(lingual.pleaseInputLabel, { label: item.label })}
-              onChange={(newValue) => {
-                updateItem(newValue, item);
-              }}
-              onBlur={() => {
-                checkItem(item);
-              }}
-              min={item.min}
-              max={item.max}
-            />
-          </div>
-        );
+        return renderNumberItem(form, item, updateItem, checkItem);
       case EMesonFieldType.Switch:
-        return (
-          <div>
-            <Switch
-              checked={form[item.name]}
-              className={styleSwitch}
-              disabled={item.disabled}
-              onChange={(value) => {
-                updateItem(value, item);
-              }}
-            />
-          </div>
-        );
+        return renderSwitchItem(form, item, updateItem, checkItem);
       case EMesonFieldType.Select:
-        let currentValue = form[item.name];
-        if (item.translateNonStringvalue && currentValue != null) {
-          currentValue = `${currentValue}`;
-        }
-        return (
-          <Select
-            value={currentValue}
-            disabled={item.disabled}
-            className={styleControlBase}
-            placeholder={item.placeholder || formatString(lingual.pleaseInputLabel, { label: item.label })}
-            onChange={(newValue) => {
-              if (item.translateNonStringvalue && newValue != null) {
-                let target = item.options.find((x) => `${x.value}` === newValue);
-                newValue = target.value;
-              }
-              updateItem(newValue, item);
-              checkItemWithValue(newValue, item);
-            }}
-            allowClear={item.allowClear}
-            onBlur={() => {
-              checkItem(item);
-            }}
-            {...item.selectProps}
-          >
-            {item.options.map((option) => {
-              let value = option.value;
-              if (item.translateNonStringvalue) {
-                value = `${value}`;
-              }
-              return (
-                <Select.Option value={value} key={option.key || value}>
-                  {option.display}
-                </Select.Option>
-              );
-            })}
-          </Select>
-        );
+        return renderSelectItem(form, item, updateItem, checkItem, checkItemWithValue);
       case EMesonFieldType.Nested:
         return renderItems(item.children);
       case EMesonFieldType.Custom:
@@ -227,19 +118,6 @@ export function ForwardForm<T = IMesonFormBase>(props: MesonFormProps<T>, ref: R
       let name: string = (item as any).name;
       let error = name != null ? errors[name] : null;
 
-      let labelNode = (
-        <div className={styleLabel}>
-          {item.required ? <RequiredMark /> : null}
-          {item.label}:
-        </div>
-      );
-
-      if (item.label == null) {
-        labelNode = <div className={styleLabel} />;
-      }
-
-      let errorNode = error != null ? <div className={styleError}>{error}</div> : null;
-
       if (item.type === EMesonFieldType.Custom) {
         let onChange = (value: any) => {
           updateItem(value, item);
@@ -249,15 +127,7 @@ export function ForwardForm<T = IMesonFormBase>(props: MesonFormProps<T>, ref: R
           checkItemWithValue(value, item);
         };
 
-        return (
-          <div key={idx} className={cx(row, styleItemRow)}>
-            {labelNode}
-            <div className={cx(flex, column, styleValueArea, item.className)} style={item.style}>
-              {item.render(form[item.name], onChange, form, onCheck)}
-              <div className={styleErrorWrapper}>{errorNode}</div>
-            </div>
-          </div>
-        );
+        return renderItemLayout(idx, item, error, item.render(form[item.name], onChange, form, onCheck));
       }
 
       if (item.type === EMesonFieldType.CustomMultiple) {
@@ -271,28 +141,12 @@ export function ForwardForm<T = IMesonFormBase>(props: MesonFormProps<T>, ref: R
 
         // errors related to multiple fields, need to extract
         let error = showErrorByNames(errors, item.names as string[]);
-        let errorNode = error != null ? <div className={styleError}>{error}</div> : null;
 
-        return (
-          <div key={idx} className={cx(row, styleItemRow)}>
-            {labelNode}
-            <div className={cx(flex, column, styleValueArea, item.className)} style={item.style}>
-              {item.renderMultiple(form, modifidForm, checkForm)}
-              <div className={styleErrorWrapper}>{errorNode}</div>
-            </div>
-          </div>
-        );
+        // notice, item CustomMultiple not handled well in layout
+        return renderItemLayout(idx, item as any, error, item.renderMultiple(form, modifidForm, checkForm));
       }
 
-      return (
-        <div key={idx} className={cx(row, styleItemRow)}>
-          {labelNode}
-          <div className={cx(flex, column, styleValueArea, item.className)} style={item.style}>
-            {renderValueItem(item)}
-            {errorNode}
-          </div>
-        </div>
-      );
+      return renderItemLayout(idx, item as any, error, renderValueItem(item));
     });
   };
 
@@ -317,6 +171,7 @@ export type FuncMesonFormForwarded<T = any> = React.ForwardRefExoticComponent<Me
 
 export let MesonForm: FuncMesonFormForwarded = React.forwardRef(ForwardForm);
 
+/** Modal binding for meson form */
 export function MesonFormModal<T>(props: {
   title: string;
   visible: boolean;
@@ -353,6 +208,7 @@ export function MesonFormModal<T>(props: {
   );
 }
 
+/** Drawer binding for meson form */
 export function MesonFormDrawer<T>(props: {
   title: string;
   visible: boolean;
@@ -395,53 +251,7 @@ let styleForm = css`
   flex: 1;
 `;
 
-let styleLabel = css`
-  color: hsla(0, 0%, 20%, 1);
-  min-width: 120px;
-  width: max-content;
-  text-align: right;
-  margin-right: 8px;
-`;
-
-let styleValueArea = css`
-  overflow: auto;
-`;
-
-let styleItemRow = css`
-  line-height: 32px;
-  margin-bottom: 16px;
-  font-size: 14px;
-`;
-
-let styleControlBase = css`
-  min-width: 180px;
-  width: 180px;
-`;
-
-let styleError = css`
-  word-break: break-all;
-  line-height: 1.5;
-  color: red;
-  padding: 4px 0px;
-`;
-
 let styleItemsContainer = css`
   overflow: auto;
   padding: 24px 16px 24px;
-`;
-
-let styleTextareaBase = css`
-  width: 240px;
-  min-width: 240px;
-`;
-
-/** 添加 wrapper 避免 error text flow 自动撑开到很大 */
-let styleErrorWrapper = css`
-  overflow: auto;
-`;
-
-let styleSwitch = css`
-  &.ant-switch {
-    margin: 4px 0;
-  }
 `;
